check update timestamp for memberlist

+++ src/dc_receive_imf.rs
@@ -1590,39 +1590,49 @@ async fn create_or_lookup_group(
 
     // add members to group/check members
     if recreate_member_list {
-        if !chat::is_contact_in_chat(context, chat_id, DC_CONTACT_ID_SELF).await {
-            // Members could have been removed while we were
-            // absent. We can't use existing member list and need to
-            // start from scratch.
-            context
-                .sql
-                .execute(
-                    "DELETE FROM chats_contacts WHERE chat_id=?;",
-                    paramsv![chat_id],
-                )
-                .await
-                .ok();
-
-            chat::add_to_chat_contacts_table(context, chat_id, DC_CONTACT_ID_SELF).await;
-        }
-        if from_id > DC_CONTACT_ID_LAST_SPECIAL
-            && !Contact::addr_equals_contact(context, &self_addr, from_id).await
-            && !chat::is_contact_in_chat(context, chat_id, from_id).await
+        if chat_id
+            .update_timestamp(context, Param::MemberListTimestamp, sent_timestamp)
+            .await?
         {
-            chat::add_to_chat_contacts_table(context, chat_id, from_id).await;
-        }
-        for &to_id in to_ids.iter() {
-            info!(context, "adding to={:?} to chat id={}", to_id, chat_id);
-            if !Contact::addr_equals_contact(context, &self_addr, to_id).await
-                && !chat::is_contact_in_chat(context, chat_id, to_id).await
-            {
-                chat::add_to_chat_contacts_table(context, chat_id, to_id).await;
+            if !chat::is_contact_in_chat(context, chat_id, DC_CONTACT_ID_SELF).await {
+                // Members could have been removed while we were
+                // absent. We can't use existing member list and need to
+                // start from scratch.
+                context
+                    .sql
+                    .execute(
+                        "DELETE FROM chats_contacts WHERE chat_id=?;",
+                        paramsv![chat_id],
+                    )
+                    .await
+                    .ok();
+
+                chat::add_to_chat_contacts_table(context, chat_id, DC_CONTACT_ID_SELF).await;
             }
+            if from_id > DC_CONTACT_ID_LAST_SPECIAL
+                && !Contact::addr_equals_contact(context, &self_addr, from_id).await
+                && !chat::is_contact_in_chat(context, chat_id, from_id).await
+            {
+                chat::add_to_chat_contacts_table(context, chat_id, from_id).await;
+            }
+            for &to_id in to_ids.iter() {
+                info!(context, "adding to={:?} to chat id={}", to_id, chat_id);
+                if !Contact::addr_equals_contact(context, &self_addr, to_id).await
+                    && !chat::is_contact_in_chat(context, chat_id, to_id).await
+                {
+                    chat::add_to_chat_contacts_table(context, chat_id, to_id).await;
+                }
+            }
+            send_EVENT_CHAT_MODIFIED = true;
         }
-        send_EVENT_CHAT_MODIFIED = true;
     } else if let Some(contact_id) = removed_id {
-        chat::remove_from_chat_contacts_table(context, chat_id, contact_id).await;
-        send_EVENT_CHAT_MODIFIED = true;
+        if chat_id
+            .update_timestamp(context, Param::MemberListTimestamp, sent_timestamp)
+            .await?
+        {
+            chat::remove_from_chat_contacts_table(context, chat_id, contact_id).await;
+            send_EVENT_CHAT_MODIFIED = true;
+        }
     }
 
     if send_EVENT_CHAT_MODIFIED {
